hide execute+apply in semantic model builder + reduce code duplication
