--- Sources/Apodini/Properties/RequestInjectable.swift
@@ -1,87 +1,0 @@
-//
-//  Request.swift
-//  Apodini
-//
-//  Created by Paul Schmiedmayer on 6/26/20.
-//
-
-import NIO
-import Vapor
-@_implementationOnly import Runtime
-
-
-protocol RequestInjectable {
-    mutating func inject(using request: Vapor.Request, with decoder: RequestInjectableDecoder?) throws
-    func accept(_ visitor: RequestInjectableVisitor)
-}
-
-extension RequestInjectable {
-    func accept(_ visitor: RequestInjectableVisitor) {
-        visitor.visit(self)
-    }
-}
-
-protocol RequestInjectableDecoder {
-    func decode<T: Decodable>(_ type: T.Type, from request: Vapor.Request) throws -> T?
-}
-
-protocol RequestInjectableVisitor {
-    func visit<Injectable: RequestInjectable>(_ requestInjectable: Injectable)
-
-    func visit<Element>(_ parameter: Parameter<Element>)
-}
-extension RequestInjectableVisitor {
-    func visit<Injectable: RequestInjectable>(_ requestInjectable: Injectable) {}
-    func visit<Element>(_ parameter: Parameter<Element>) {}
-}
-
-
-private func extractRequestInjectables<Element>(from subject: Element) -> [String: RequestInjectable] {
-    var result: [String: RequestInjectable] = [:]
-    
-    execute({ (injectable: RequestInjectable, label: String) in
-        result[label] = injectable
-    }, on: subject)
-    
-    return result
-}
-
-extension Component {
-    func extractRequestInjectables() -> [String: RequestInjectable] {
-        Apodini.extractRequestInjectables(from: self)
-    }
-}
-extension AnyResponseTransformer {
-    func extractRequestInjectables() -> [String: RequestInjectable] {
-        Apodini.extractRequestInjectables(from: self)
-    }
-}
-
-extension Vapor.Request {
-    func enterRequestContext<E, R>(with element: E, using decoder: RequestInjectableDecoder? = nil, executing method: (E) -> EventLoopFuture<R>)
-                    -> EventLoopFuture<R> {
-        var element = element
-        inject(in: &element, using: decoder)
-
-        return method(element)
-    }
-
-    func enterRequestContext<E, R>(with element: E, using decoder: RequestInjectableDecoder? = nil, executing method: (E) -> R) -> R {
-        var element = element
-        inject(in: &element, using: decoder)
-        return method(element)
-    }
-    
-    private func inject<E>(in element: inout E, using decoder: RequestInjectableDecoder? = nil) {
-        // Inject all properties that can be injected using RequestInjectable
-        let elem = element
-        
-        apply({ (requestInjectable: inout RequestInjectable) in
-            do {
-                try requestInjectable.inject(using: self, with: decoder)
-            } catch {
-                fatalError("Injecting into element \(elem) failed.")
-            }
-        }, to: &element)
-    }
-}
