+++ Sources/Apodini/Semantic Model Builder/RequestInjectable.swift
@@ -1,0 +1,47 @@
+//
+//  RequestInjectable.swift
+//  Apodini
+//
+//  Created by Paul Schmiedmayer on 6/26/20.
+//
+
+import NIO
+import Vapor
+@_implementationOnly import Runtime
+
+
+protocol RequestInjectable {
+    mutating func inject(using request: Vapor.Request, with decoder: RequestInjectableDecoder?) throws
+    func accept(_ visitor: RequestInjectableVisitor)
+}
+
+extension RequestInjectable {
+    func accept(_ visitor: RequestInjectableVisitor) {
+        visitor.visit(self)
+    }
+}
+
+protocol RequestInjectableDecoder {
+    func decode<T: Decodable>(_ type: T.Type, from request: Vapor.Request) throws -> T?
+}
+
+protocol RequestInjectableVisitor {
+    func visit<Injectable: RequestInjectable>(_ requestInjectable: Injectable)
+
+    func visit<Element>(_ parameter: Parameter<Element>)
+}
+extension RequestInjectableVisitor {
+    func visit<Injectable: RequestInjectable>(_ requestInjectable: Injectable) {}
+    func visit<Element>(_ parameter: Parameter<Element>) {}
+}
+
+extension Component {
+    func extractRequestInjectables() -> [String: RequestInjectable] {
+        Apodini.extractRequestInjectables(from: self)
+    }
+}
+extension AnyResponseTransformer {
+    func extractRequestInjectables() -> [String: RequestInjectable] {
+        Apodini.extractRequestInjectables(from: self)
+    }
+}
